:recycle: 4 decimals precision in Liquidator

+++ test/utils/Events.sol
@@ -55,7 +55,7 @@ abstract contract Events {
     event WeightsSet(uint16 initiatorRewardWeight, uint16 penaltyWeight, uint16 closingRewardWeight);
     event AuctionCurveParametersSet(uint64 base, uint32 cutoffTime);
     event StartPriceMultiplierSet(uint16 startPriceMultiplier);
-    event MinimumPriceMultiplierSet(uint8 minPriceMultiplier);
+    event MinimumPriceMultiplierSet(uint16 minPriceMultiplier);
     event AuctionFinished(
         address indexed account, address indexed creditor, uint128 startDebt, uint128 totalBids, uint128 badDebt
     );
